moved ngen_main to ngen_routing dir and cleaned up other ngen related functions
