+++ src/ngen_routing/src/ngen_main.py
@@ -1,0 +1,23 @@
+import argparse
+import time
+from datetime import datetime
+from collections import defaultdict
+import pathlib
+import pandas as pd
+import sys
+import os
+
+if not hasattr(sys, 'argv'):
+    sys.argv  = ['']
+
+
+def set_paths(root_path):
+
+    return
+
+
+def ngen_main(argv):
+    from nwm_routing.__main__ import main_v02
+    main_v02(argv)
+
+
